+++ src/graphics/color/colorfunc.js
@@ -60,7 +60,7 @@ export function hexToArray (hex, alpha) {
   ] : null
 }
 
-function componentToHex (c) {
+export function componentToHex (c) {
   var hex = c.toString(16)
   return hex.length === 1 ? '0' + hex : hex
 }
