+++ src/map/layer/webgl/temporalHeatmap-layer/temporalHeatmap-animation-vertex.glsl.js
